backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
aiofiles>=24.1.0
Pillow>=11.0.0
pdf2image>=1.17.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
